Fix data type conversion when creating a clone table in memory. Column loader now accepts an argument describing what data type is used to express each field.
Better handling of JdbcSchema when the list of tables in schema has been pre-populated but the columns of a particular table have not.

Make timeZone available for queries that populate in-memory tables.

Refactor and add tests for code relating to Mapping.

Fix how strings are escaped in JsonBuilder.

By default, test suite runs against hsqldb but you can pass '-Doptiq.test.db=mysql' to override.