Add a way to specify list of TraitDefs that will be used by planner when create planner.
Allow PlannerImpl to use a user-specified RelTraitDefs, in stead of default RelTraitDefs.

Add a new getPlanner to FrameWorks to specify the list of RelTraitDef.

Add a unit test case to test planner with explicitly specified list of RelTraitDefs.

Minor change to unit testcase for TraitDef.

Add comment to Framework's new method.