Fix https://github.com/julianhyde/optiq/issues/229, "Join between different JDBC schemas not implementable".
A JdbcConvention was registered when the first JdbcTableScan was registered, but the other instance of JdbcConvention (for the other data source) was not, because it was not the first instance of class JdbcTableScan. The solution was to add a RelTrait.register(RelOptPlanner) method. It is called for each trait present in the initial RelNode tree.

Test case contributed by Bruno Dumon.

This is a breaking change. Classes that implement RelTrait will need to add a 'register(RelOptPlanner)' method.