Fix TestTraitPropagation.withoutHack (withoutHack now succeeds, withHack now fails).
Deduce the collations of a RelSubset by looking at its traits. (For other kinds of RelNode that would cause a cycle.)

Remove AbstractConverters again (Jacques had restored them, but performance was terrible.)

When creating a RelSubset make sure that its traits are simple. (Occurs when creating a subset to convert an existing rel that has multiple collations to a different calling convention. The subset can only have one collation.)

Add a version of RelOptRule.convert that takes a single trait, to avoid simplifying other traits of that RelNode. Many current calls to convert(RelNode, RelTraitSet) should probably use this method.

Add a short-cut to RelTraitSet.replace for the case where the new trait is canonized and the same as the old.

In TestTraitPropagation, use SortRemoveRule; implement RelOptTable.unwrap and Table.getStatistic and use EnumerableTableScan.create, so that table scan is sorted; make PhysSortRule a ConvertRule, so that its input ends up in the same RelSet; fix a typo giving PhysTable the wrong convention.

Call simplify in RelOptRule.convert rather than VolcanoPlanner.