[CALCITE-1684] Change default precision of VARCHAR and VARBINARY from 1 to "unspecified" (Kevin Liew)
If you concatenate two strings and the sum of their precisions is
greater than the maximum precision (65,536 by default on Calcite, but
can be overriden in RelDataTypeSystem) the result precision will be
unspecified.

When casting from unspecified precision to a particular precision,
may need to truncate values.

Some unrelated USING tests, and fix typos in method names.

Close apache/calcite#394